requirements.txt
opencv-python~=4.11.0.86
numpy~=1.26.4
matplotlib~=3.8.4
pillow~=10.3.0
pygrabber
PyYAML~=6.0.2
torch~=2.5.1+cu121
scikit-image~=0.25.2
torchvision~=0.20.1+cu121